+++ s/compat/types.ts
@@ -9,6 +9,6 @@ export type Compat = {
 	getSelf(): PostableChannel
 	guessOptimalThreadCount(): number
 	loadWorker(url: string | URL, name: string | undefined): CompatWorker
-	loadWasm(url: string | URL): Promise<WebAssembly.WebAssemblyInstantiatedSource>
+	loadWasm(url: string | URL): Promise<WebAssembly.WebAssemblyInstantiatedSource | WebAssembly.Instance>
 }