+++ s/comrade.ts
@@ -1,12 +1,12 @@
 
-import {makeMock, Rig, Tap} from "@e280/renraku"
+import {makeMock, Tap} from "@e280/renraku"
 
 import {worker} from "./parts/worker.js"
 import {Compat} from "./compat/types.js"
 import {Thread} from "./parts/thread.js"
+import {shells} from "./parts/shells.js"
 import {Cluster} from "./parts/cluster.js"
 import {defaultTap} from "./parts/default-tap.js"
-import {HostShell, WorkShell} from "./parts/shells.js"
 import {ClusterOptions, Mocks, Schematic, SetupHost, SetupWork, ThreadOptions, WorkerOpts} from "./parts/types.js"
 
 export const setupComrade = (compat: Compat) => ({
@@ -28,11 +28,11 @@ export const setupComrade = (compat: Compat) => ({
 
 		const {setupWork, setupHost, tap = defaultTap} = options
 
-		const hostShell = new HostShell<S>()
-		const workShell = new WorkShell<S>()
+		const hostShell = shells.mock.host<S>()
+		const workShell = shells.mock.work<S>()
 
-		workShell.work = makeMock({tap, fns: setupWork(hostShell, new Rig())})
-		hostShell.host = makeMock({tap, fns: setupHost(workShell, new Rig())})
+		workShell.work = makeMock({tap, fns: setupWork(hostShell)})
+		hostShell.host = makeMock({tap, fns: setupHost(workShell)})
 
 		return {
 			workShell,
@@ -43,17 +43,17 @@ export const setupComrade = (compat: Compat) => ({
 	},
 
 	mockWork<S extends Schematic>(setupWork: SetupWork<S>, tap: Tap = defaultTap) {
-		const hostShell = new HostShell<S>()
-		const workShell = new WorkShell<S>()
+		const hostShell = shells.mock.host<S>()
+		const workShell = shells.mock.work<S>()
 
-		workShell.work = makeMock({tap, fns: setupWork(hostShell, new Rig())})
+		workShell.work = makeMock({tap, fns: setupWork(hostShell)})
 
 		return {
 			workShell,
 			hostShell,
 			work: workShell.work,
 			mockHost: (setupHost: SetupHost<S>): Mocks<S> => {
-				hostShell.host = makeMock({tap, fns: setupHost(workShell, new Rig())})
+				hostShell.host = makeMock({tap, fns: setupHost(workShell)})
 				return {
 					workShell,
 					hostShell,